abstracted component
