+++ native-app/components/MachineState.tsx
@@ -1,0 +1,129 @@
+import { Link } from "expo-router";
+import { Button, StyleSheet } from "react-native";
+import Loading from "../components/Loading";
+import { MachineScore } from "../components/MachineScore";
+import { PartsOfMachine } from "../components/PartsOfMachine";
+import { Text, View } from "../components/Themed";
+
+type Props = {
+  machineValues?: MachineValues;
+  machineHealth?: MachineHealthResponseBody;
+  isLoading?: boolean;
+  calculateHealth(): void;
+  resetMachineValues(): void;
+};
+
+const MachineState: React.FC<Props> = (props: Props) => {
+  const {
+    machineValues,
+    machineHealth,
+    calculateHealth,
+    isLoading,
+    resetMachineValues,
+  } = props;
+
+  return (
+    <View style={styles.container}>
+      <View style={styles.separator} />
+      {isLoading && <Loading />}
+
+      {!machineValues && (
+        <Link href="/two" style={styles.link}>
+          <Text style={styles.linkText}>
+            Please log a part to check machine health
+          </Text>
+        </Link>
+      )}
+
+      {!!machineValues && (
+        <>
+          <PartsOfMachine
+            machineName={"Welding Robot"}
+            parts={machineValues.machines.weldingRobot}
+          />
+          <PartsOfMachine
+            machineName={"Assembly Line"}
+            parts={machineValues.machines.assemblyLine}
+          />
+          <PartsOfMachine
+            machineName={"Painting Station"}
+            parts={machineValues.machines.paintingStation}
+          />
+          <PartsOfMachine
+            machineName={"Quality Control Station"}
+            parts={machineValues.machines.qualityControlStation}
+          />
+          <View
+            style={styles.separator}
+            lightColor="#eee"
+            darkColor="rgba(255,255,255,0.1)"
+          />
+          <Text style={styles.title}>Factory Health Score</Text>
+          <Text style={styles.text}>
+            {machineHealth && machineHealth.factory
+              ? machineHealth.factory
+              : "Not yet calculated"}
+          </Text>
+          {machineHealth && machineHealth.machineScores && (
+            <>
+              <Text style={styles.title2}>Machine Health Scores</Text>
+              {Object.keys(machineHealth.machineScores).map((key) => (
+                <MachineScore
+                  key={key}
+                  machineName={key}
+                  score={machineHealth.machineScores[key]}
+                />
+              ))}
+            </>
+          )}
+        </>
+      )}
+      <View
+        style={styles.separator}
+        lightColor="#eee"
+        darkColor="rgba(255,255,255,0.1)"
+      />
+      <Button title="Calculate Health" onPress={calculateHealth} />
+      <View style={styles.resetButton}>
+        <Button
+          title="Reset Machine Data"
+          onPress={resetMachineValues}
+          color="#FF0000"
+        />
+      </View>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+  },
+  title2: {
+    fontSize: 17,
+    fontWeight: "bold",
+  },
+  separator: {
+    marginVertical: 20,
+    height: 1,
+    width: "80%",
+  },
+  text: {},
+  link: {
+    paddingBottom: 15,
+  },
+  linkText: {
+    fontSize: 14,
+    color: "#2e78b7",
+  },
+  resetButton: {
+    marginTop: 10,
+  },
+});
+
+export default MachineState;
